// TypeScript/src/client/message_protocol.ts
type ServerData = Record<string, unknown>;
// ...
export class Position {
  public constructor(
    public readonly X: number,
    public readonly Y: number,
  ) {}
}
// ...
export class VisibleStructure {
  public Id = 0;
  public Type = "Extractor";
  public OwnerTeamId = 0;
  public OwnerBaseId = 0;
  public IsAlly = false;
  public PvpActivated = false;
  public Position = new Position(0, 0);
  public Width = 1;
  public Height = 1;
  public Hp = 0;
  public MaxHp = 0;
  public PowerOn = false;
  public IsActive = false;
  public VisionRadius = 0;

  public static fromServer(data: ServerData): VisibleStructure {
    const structure = new VisibleStructure();
    structure.Id = asNumber(data.id);
    structure.Type = String(getValue(data, "type", "Extractor"));
    structure.OwnerTeamId = asNumber(data.ownerTeamId);
    structure.OwnerBaseId = asNumber(data.ownerBaseId);
    structure.IsAlly = asBoolean(data.isAlly);
    structure.PvpActivated = asBoolean(data.pvpActivated);
    structure.Position = positionFrom(data.position);
    structure.Width = asNumber(data.width, 1);
    structure.Height = asNumber(data.height, 1);
    structure.Hp = asNumber(data.hp);
    structure.MaxHp = asNumber(data.maxHp);
    structure.PowerOn = asBoolean(data.powerOn);
    structure.IsActive = asBoolean(data.isActive);
    structure.VisionRadius = asNumber(data.visionRadius);
    return structure;
  }
}
// ...
export abstract class ActionBase {
  public constructor(public readonly Type: string) {}

  public toServerPayload(): ServerData {
    return { type: this.Type };
  }
}

export class MoveAction extends ActionBase {
  public constructor(public readonly NewPosition: Position) {
    super("Move");
  }

  public override toServerPayload(): ServerData {
    return { type: this.Type, newPosition: positionPayload(this.NewPosition) };
  }
}

export class GatherNodeAction extends ActionBase {
  public constructor(public readonly GatherPosition: Position) {
    super("Gather");
  }

  public override toServerPayload(): ServerData {
    return { type: this.Type, gatherPosition: positionPayload(this.GatherPosition) };
  }
}

export class AttackAction extends ActionBase {
  public constructor(public readonly TargetPosition: Position) {
    super("Attack");
  }

  public override toServerPayload(): ServerData {
    return { type: this.Type, targetPosition: positionPayload(this.TargetPosition) };
  }
}